attacksim/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Security-Too==5.4.3
Flask-WTF>=1.1.2
Flask-Mail==0.9.1
Flask-Limiter==3.5.0
SQLAlchemy==2.0.36
Werkzeug==2.3.7
Jinja2==3.1.2
WTForms==3.0.1
email-validator==2.0.0
bcrypt==4.0.1
PyJWT==2.8.0
celery==5.3.1
redis==5.0.1
gunicorn==21.2.0
pytest==7.4.2
black==23.9.1
flake8==6.1.0
python-dotenv==1.0.0
cryptography==41.0.7
Authlib==1.3.0
requests==2.31.0
Flask-Babel==4.0.0 